requirements.txt
cohere==2.5.0
numpy==1.19.4
pandas==1.2.2
python-dotenv==0.21.0
